project 10- time related
Download the code from this chapter from https: // thinkpython. com/ code/ Time2. py . Change the attributes of Time to be a single integer representing seconds since midnight. Then modify the methods (and the function int__to__time) to work with the new implementation. You should not have to modify the test code in main. When you are done, the output should be the same as before.